Update magic_link_email.php
The `route_to()` returns just the relative url which is only useful within the application. To make this work outside the application, the url has to be an absolute url, which `url_to()` does perfectly (even with named routes) as it does a `route_to()` behind the scenes and attaches it to the `site_url()`.