Update Email2FA.php
When a user successfully logs in using the 2FA method, I believe the `didLogin` event should also be fired just like it does in the login without the 2FA action. 

This should be so, because, say for example, a developer is trying to listen to that event, with the way it is currently, it will only work when the 2FA action is not set.